app/requirements.txt
langchain
streamlit
langchain-community
sentence-transformers
pinecone-client
openai
torch
transformers
nltk
streamlit-chat